add migration

+++ projects/todo/internal/todo/types.go
@@ -1,8 +1,8 @@
 package todo
 
 type Todo struct {
-	ID     int    `json:"id"`
-	UserID int    `json:"userId"`
+	ID     int64  `json:"id"`
+	UserID int64  `json:"userId"`
 	Name   string `json:"name"`
 	Done   bool   `json:"done"`
 }
@@ -10,5 +10,5 @@ type Todo struct {
 type NewTodoDTO struct {
 	Name   string `json:"name" validate:""`
 	Done   bool   `json:"done" validate:""`
-	UserID int
+	UserID int64
 }
